feat: login route. hide all app UI until logged in
