+++ client/assets/css/tailwind.css
@@ -22,3 +22,8 @@
     z-index: 1000;
   }
 }
+
+#__nuxt {
+  width: 100%;
+  height: 100%;
+}
